requirements.txt
aiohttp==3.1.3
aniso8601==3.0.0
async-timeout==2.0.1
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-JSON==0.3.2
Flask-JWT-Extended==3.8.1
flask-mongoengine==0.9.5
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-WTF==0.14.2
idna==2.6
idna-ssl==1.0.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mock==2.0.0
mongoengine==0.15.0
mongomock==3.9.0
more-itertools==4.1.0
multidict==4.1.0
pbr==4.0.2
pluggy==0.6.0
py==1.5.3
PyJWT==1.6.1
pymongo==3.6.1
pytest==3.5.0
pytz==2018.4
requests==2.18.4
sentinels==1.0.0
six==1.11.0
telepot==12.6
urllib3==1.22
uWSGI==2.0.17
Werkzeug==0.14.1
WTForms==2.1
yarl==1.1.1
